+++ extralit-frontend/v1/domain/services/IFileMatchingService.ts
@@ -1,4 +1,4 @@
-import type { ParsedEntry } from "./IFileService";
+import type { ParsedEntry } from "./IFileParsingService";
 
 export interface FileMatchResult {
   file: File;
